+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 0.10.8-dev
+ - introduce `--report-file` (and `GooseDefault::ReportFile`) to optionally generate an HTML report when the load test completes
 
 ## 0.10.7 Nov 16, 2020
  - account for time spent doing things other than sleeping, maintaining more consistency when displaying statistics and shutting down